server/requirements/requirements.txt
fastapi
psycopg2
psycopg2-binary
passlib
python-jose